+++ scripts/MAIN.py
@@ -140,7 +140,6 @@ def MAIN(progress_callback=None, total_items_callback=None, prefix_callback=None
 
         check_stop_flag()
 
-        first_run = False
         update = False
 
         # STEP 4: cleaning spectrums (Multithreaded)
@@ -148,7 +147,7 @@ def MAIN(progress_callback=None, total_items_callback=None, prefix_callback=None
         if step_callback:
             step_callback("-- CHECKING FOR UPDATES --")
         time.sleep(0.01)
-        spectrum_list, update_temp, first_run_temp = check_for_update_processing(spectrum_list, output_directory,
+        spectrum_list, update_temp = check_for_update_processing(spectrum_list, output_directory,
                                                                                  progress_callback=progress_callback,
                                                                                  total_items_callback=total_items_callback,
                                                                                  prefix_callback=prefix_callback,
@@ -161,8 +160,10 @@ def MAIN(progress_callback=None, total_items_callback=None, prefix_callback=None
 
             if update_temp:
                 update = True
-            if first_run_temp:
-                first_run = True
+
+            if parameters_dict['reset_updates'] == 1.0:
+                update = False
+
             time.sleep(0.01)
             if step_callback:
                 step_callback("-- CLEANING SPECTRUMS --")
@@ -311,7 +312,7 @@ def MAIN(progress_callback=None, total_items_callback=None, prefix_callback=None
                     step_callback("--  WRITING CSV --")
                 time.sleep(0.01)
                 writting_csv(POS_LC_df, POS_GC_df, NEG_LC_df, NEG_GC_df, POS_LC_In_Silico_df, POS_GC_In_Silico_df,
-                             NEG_LC_In_Silico_df, NEG_GC_In_Silico_df, first_run, output_directory, update,
+                             NEG_LC_In_Silico_df, NEG_GC_In_Silico_df, output_directory, update,
                              progress_callback=progress_callback, total_items_callback=total_items_callback,
                              prefix_callback=prefix_callback, item_type_callback=item_type_callback)