+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+- **01_10_2025**:
+  - fixing missing headers issue when writing a previously not existing csv file.
+
 
 - **26_09_2025**:
   - remove peaks with intensity <= 0.